requirements.txt
ultralytics>=8.4.18
opencv-python-headless>=4.8.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pyyaml>=6.0
jinja2>=3.1.0
python-multipart>=0.0.6
Pillow>=10.0.0
